% Generated by roxygen2 (4.0.1): do not edit by hand
\name{assayColorMap<-}
\alias{assayColorMap<-}
\title{Set assayColorMap for object of EigensystemPlotParam}
\usage{
assayColorMap(x) <- value
}
\arguments{
\item{x}{object of class Eigensystem}

\item{value}{new value for slot assayColorMap}
}
\value{
object of class Eigensystem
}
\description{
Set assayColorMap for object of EigensystemPlotParam
}
\details{
Set assayColorMap for object of EigensystemPlotParam
}
\author{
Anneleen Daemen \email{daemen.anneleen@gene.com}, Matthew Brauer \email{brauer.matthew@gene.com}
}
\seealso{
EigensystemPlotParam-class
}

