% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{biosvd}
\alias{biosvd}
\alias{biosvd-package}
\title{Package for high-throughput data processing, outlier detection, noise removal and dynamic modeling}
\description{
Biosvd package for high-throughput data processing, outlier detection, noise removal and dynamic modeling
}
\details{
This package allows for high-throughput data processing, outlier detection, noise removal and dynamic modeling,
based on the framework of Singular Value Decomposition. It provides the user with summary graphs and an
interactive html report.
}
\examples{
example( report )
}
\author{
Anneleen Daemen \email{daemen.anneleen@gene.com}, Matthew Brauer \email{brauer.matthew@gene.com}
}
\keyword{package}

