% Generated by roxygen2 (4.0.1): do not edit by hand
\name{contrast<-}
\alias{contrast<-}
\title{Set contrast for object of EigensystemPlotParam}
\usage{
contrast(x) <- value
}
\arguments{
\item{x}{object of class Eigensystem}

\item{value}{new value for slot contrast}
}
\value{
object of class Eigensystem
}
\description{
Set contrast for object of EigensystemPlotParam
}
\details{
Set contrast for object of EigensystemPlotParam
}
\author{
Anneleen Daemen \email{daemen.anneleen@gene.com}, Matthew Brauer \email{brauer.matthew@gene.com}
}
\seealso{
EigensystemPlotParam-class
}

