% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{EigensystemPlotParam-class}
\alias{EigensystemPlotParam}
\alias{EigensystemPlotParam-class}
\alias{assayColorMap}
\alias{contrast}
\alias{featureColorMap}
\alias{figure}
\alias{filenames}
\alias{negativeValues}
\alias{palette}
\alias{path}
\alias{plots}
\alias{prefix}
\alias{show,EigensystemPlotParam-method}
\alias{whichAssays}
\alias{whichEigenassays}
\alias{whichEigenfeatures}
\alias{whichFeatures}
\alias{whichPolarAxes}
\title{The EigensystemPlotParam class}
\description{
\code{EigensystemPlotParam} is a list-based class for storing the parameters needed to specify plot features used by \code{link{plot,plot-method}} in plotting Eigensystem diagnostics, projections and transformations.
}
\details{
The \code{EigensystemPlotParam} class stores the list of desired plots, color palettes and keys, figure file and directory names and other necessary parameters.
Data in the \code{EigensystemPlotParam} may be organized into slots:
\sQuote{plots}, \sQuote{palette}, \sQuote{whichAssays}, \sQuote{whichFeatures},
\sQuote{whichEigenassays}, \sQuote{whichEigenfeatures}, \sQuote{whichPolarAxes},
\sQuote{assayColorMap}, \sQuote{featureColorMap}, \sQuote{contrast},
\sQuote{negativeValues}, \sQuote{path}, \sQuote{prefix},
\sQuote{filename}, \sQuote{figure},
brief descriptions of which follow.
}
\section{Slots}{

\code{EigensystemPlotParam} objects contain the following slots
\describe{
	\item{\code{plots}:}{\code{character vector} indicating one or more plot choices from: "eigenfeatureHeatmap", "eigenassayHeatmap", "sortedHeatmap",
          "fraction","scree","zoomedFraction", "lines", "allLines", "eigenfeaturePolar", "eigenassayPolar". Defaults to all. }
	\item{\code{palette}:}{\code{function} defining the palette to be used for heatmaps. Default is a Blue-Yellow color ramp. }
	\item{\code{whichAssays}:}{\code{numeric vector} listing which assays are to be included in the plot(s). Default is all. }
	\item{\code{whichFeatures}:}{\code{numeric vector} which features to include in plots. Default is all. }
	\item{\code{whichEigenassays}:}{\code{numeric vector}  which eigenassays to include in plots. }
	\item{\code{whichEigenfeatures}:}{\code{numeric vector} which eigenfeatures to include in plots. Defaults to first four. }
    \item{\code{whichPolarAxes}:}{\code{numeric vector} which two eigenassays/eigenfeatures to include in polar plots. Default is first two. }
	\item{\code{assayColorMap}:}{ assayColorMap and featureColorMap are optional \code{list}s of colors corresponding to the
          levels of these annotations for assays and features. The elements of each list are named by the annotation column they
          correspond to. Each element is itself a named vector of colors, named by the level of the annotation it reflects (see example). }
	\item{\code{featureColorMap}:}{\code{list} }
	\item{\code{contrast}:}{\code{numeric} value specifying the contrast to use in heatmaps. }
	\item{\code{negativeValues}:}{\code{logical} indicating whether scaling of values for heatmap should result in range that includes negative values. Default is TRUE.}
	\item{\code{path}:}{\code{character} specifying the path of files for figure output. Default is current working directory.}
	\item{\code{prefix}:}{\code{character} specifying an optional prefix to add to filenames. Default is biosvd.}
	\item{\code{filenames}:}{\code{character} optional name for the files containing the plots. Names correspond to elements of 'plots' vector.}
	\item{\code{figure}:}{\code{logical} specifying whether to ouptut plots into files. Default is FALSE. }
}
}

\section{Accessors}{

\describe{
   \item{}{\code{plots(x)}, \code{plots(x) <- value}}
   \item{}{\code{palette(x)}, \code{palette(x) <- value}}
   \item{}{\code{whichAssays(x)}, \code{whichAssays(x) <- value}}
   \item{}{\code{whichFeatures(x)}, \code{whichFeatures(x) <- value}}
   \item{}{\code{whichEigenassays(x)}, \code{whichEigenassays(x) <- value}}
   \item{}{\code{whichEigenfeatures(x)}, \code{whichEigenfeatures(x) <- value}}
   \item{}{\code{whichPolarAxes(x)}, \code{whichPolarAxes(x) <- value}}
   \item{}{\code{assayColorMap(x)}, \code{assayColorMap(x) <- value}}
   \item{}{\code{featureColorMap(x)}, \code{featureColorMap(x) <- value}}
   \item{}{\code{contrast(x)}, \code{contrast(x) <- value}}
   \item{}{\code{negativeValues(x)}, \code{negativeValues(x) <- value}}
   \item{}{\code{path(x)}, \code{path(x) <- value}}
   \item{}{\code{prefix(x)}, \code{prefix(x) <- value}}
   \item{}{\code{filenames(x)}, \code{filenames(x) <- value}}
   \item{}{\code{figure(x)}, \code{figure(x) <- value}}
 }
}
\examples{
data(YeastData_alpha)

params <- new("EigensystemPlotParam")
cellcycle.col.map <- c("orange2", "darkgreen", "blue2", "magenta2", "red2")
names(cellcycle.col.map) <- c("S", "G2", "M", "M/G1", "G1")
assayColorMap(params) <- list(Cell.cycle.stage=cellcycle.col.map)
featureColorMap(params) <- list(Cell.cycle.stage=NA)
}
\author{
Anneleen Daemen \email{daemen.anneleen@gene.com}, Matthew Brauer \email{brauer.matthew@gene.com}
}
\references{
Alter O, Brown PO and Botstein D. Singular value decomposition for genome-wide expression data processing and modeling. Proc Natl Acad Sci U.S.A. 97(18), 10101-10106 (2000).
}
\keyword{classes}
\keyword{methods}

