% Generated by roxygen2 (4.0.1): do not edit by hand
\name{featureColorMap<-}
\alias{featureColorMap<-}
\title{Set featureColorMap for object of EigensystemPlotParam}
\usage{
featureColorMap(x) <- value
}
\arguments{
\item{x}{object of class Eigensystem}

\item{value}{new value for slot featureColorMap}
}
\value{
object of class Eigensystem
}
\description{
Set featureColorMap for object of EigensystemPlotParam
}
\details{
Set featureColorMap for object of EigensystemPlotParam
}
\author{
Anneleen Daemen \email{daemen.anneleen@gene.com}, Matthew Brauer \email{brauer.matthew@gene.com}
}
\seealso{
EigensystemPlotParam-class
}

