% Generated by roxygen2 (4.0.1): do not edit by hand
\name{figure<-}
\alias{figure<-}
\title{Set figure for object of EigensystemPlotParam}
\usage{
figure(x) <- value
}
\arguments{
\item{x}{object of class Eigensystem}

\item{value}{new value for slot figure}
}
\value{
object of class Eigensystem
}
\description{
Set figure for object of EigensystemPlotParam
}
\details{
Set figure for object of EigensystemPlotParam
}
\author{
Anneleen Daemen \email{daemen.anneleen@gene.com}, Matthew Brauer \email{brauer.matthew@gene.com}
}
\seealso{
EigensystemPlotParam-class
}

