% Generated by roxygen2 (4.0.1): do not edit by hand
\name{filenames<-}
\alias{filenames<-}
\title{Set filenames for object of EigensystemPlotParam}
\usage{
filenames(x) <- value
}
\arguments{
\item{x}{object of class Eigensystem}

\item{value}{new value for slot filenames}
}
\value{
object of class Eigensystem
}
\description{
Set filenames for object of EigensystemPlotParam
}
\details{
Set filenames for object of EigensystemPlotParam
}
\author{
Anneleen Daemen \email{daemen.anneleen@gene.com}, Matthew Brauer \email{brauer.matthew@gene.com}
}
\seealso{
EigensystemPlotParam-class
}

