% Generated by roxygen2 (4.0.1): do not edit by hand
\name{negativeValues<-}
\alias{negativeValues<-}
\title{Set negativeValues for object of EigensystemPlotParam}
\usage{
negativeValues(x) <- value
}
\arguments{
\item{x}{object of class Eigensystem}

\item{value}{new value for slot negativeValues}
}
\value{
object of class Eigensystem
}
\description{
Set negativeValues for object of EigensystemPlotParam
}
\details{
Set negativeValues for object of EigensystemPlotParam
}
\author{
Anneleen Daemen \email{daemen.anneleen@gene.com}, Matthew Brauer \email{brauer.matthew@gene.com}
}
\seealso{
EigensystemPlotParam-class
}

