% Generated by roxygen2 (4.0.1): do not edit by hand
\name{palette<-}
\alias{palette<-}
\title{Set palette for object of EigensystemPlotParam}
\usage{
palette(x) <- value
}
\arguments{
\item{x}{object of class Eigensystem}

\item{value}{new value for slot palette}
}
\value{
object of class Eigensystem
}
\description{
Set palette for object of EigensystemPlotParam
}
\details{
Set palette for object of EigensystemPlotParam
}
\author{
Anneleen Daemen \email{daemen.anneleen@gene.com}, Matthew Brauer \email{brauer.matthew@gene.com}
}
\seealso{
EigensystemPlotParam-class
}

