% Generated by roxygen2 (4.0.1): do not edit by hand
\name{path<-}
\alias{path<-}
\title{Set path for object of EigensystemPlotParam}
\usage{
path(x) <- value
}
\arguments{
\item{x}{object of class Eigensystem}

\item{value}{new value for slot path}
}
\value{
object of class Eigensystem
}
\description{
Set path for object of EigensystemPlotParam
}
\details{
Set path for object of EigensystemPlotParam
}
\author{
Anneleen Daemen \email{daemen.anneleen@gene.com}, Matthew Brauer \email{brauer.matthew@gene.com}
}
\seealso{
EigensystemPlotParam-class
}

