% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plots<-}
\alias{plots<-}
\title{Set plots for object of EigensystemPlotParam}
\usage{
plots(x) <- value
}
\arguments{
\item{x}{object of class Eigensystem}

\item{value}{new value for slot plots}
}
\value{
object of class Eigensystem
}
\description{
Set plots for object of EigensystemPlotParam
}
\details{
Set plots for object of EigensystemPlotParam
}
\author{
Anneleen Daemen \email{daemen.anneleen@gene.com}, Matthew Brauer \email{brauer.matthew@gene.com}
}
\seealso{
EigensystemPlotParam-class
}

