% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{project,Eigensystem-method}
\alias{project}
\alias{project,Eigensystem-method}
\title{Projects the data onto one or two eigenfeatures and eigenassays}
\usage{
\S4method{project}{Eigensystem}(x, axes = c(2, 1), type = "features")
}
\arguments{
\item{x}{object of class Eigensystem}

\item{axes}{numerical vector specifying eigenfeatures to project onto (default c(2,1))}

\item{type}{string specifying the dimensions to return the coordinates for, either features or assays (default features)}
}
\value{
Object of class eigensystem
}
\description{
Returns the rectangular and polar coordinates of the eigensystem projection onto one or two eigenfeatures and eigenassays
}
\details{
The function allows the sorting of the data according to one or two specified eigenfeatures and eigenassays. This gives a global picture of the dynamics of expression/intensities, in which individual features and assays are classified in groups of similar regulation and function or similar cellular state and biological phenotype.
}
\section{Methods}{

\describe{
\item{\code{signature(x = "Eigensystem")}}{
}
}
}
\examples{
## Data obtained from http://genomics-pubs.princeton.edu/StarvationMetabolomics/Download.shtm
data(StarvationData)

## Computes the eigensystem for the actual data
eigensystem <- compute(StarvationData)
## Excludes the eigenfeatures representing steady-state intensity
eigensystem <- exclude(eigensystem)

## Find the projection of the data onto eigenfeature 1 and 2
projection <- project(eigensystem)
## Project the data onto eigenfeature 3 and 4
projection <- project(eigensystem, axes=c(4,3))
}
\author{
Anneleen Daemen \email{daemen.anneleen@gene.com}, Matthew Brauer \email{brauer.matthew@gene.com}
}
\references{
Alter O, Brown PO and Botstein D. Singular value decomposition for genome-wide expression data processing and modeling. Proc Natl Acad Sci U.S.A. 97(18), 10101-10106 (2000).
}

