% Generated by roxygen2 (4.0.1): do not edit by hand
\name{whichAssays<-}
\alias{whichAssays<-}
\title{Set whichAssays for object of EigensystemPlotParam}
\usage{
whichAssays(x) <- value
}
\arguments{
\item{x}{object of class Eigensystem}

\item{value}{new value for slot whichAssays}
}
\value{
object of class Eigensystem
}
\description{
Set whichAssays for object of EigensystemPlotParam
}
\details{
Set whichAssays for object of EigensystemPlotParam
}
\author{
Anneleen Daemen \email{daemen.anneleen@gene.com}, Matthew Brauer \email{brauer.matthew@gene.com}
}
\seealso{
EigensystemPlotParam-class
}

