% Generated by roxygen2 (4.0.1): do not edit by hand
\name{whichEigenfeatures<-}
\alias{whichEigenfeatures<-}
\title{Set whichEigenfeatures for object of EigensystemPlotParam}
\usage{
whichEigenfeatures(x) <- value
}
\arguments{
\item{x}{object of class Eigensystem}

\item{value}{new value for slot whichEigenfeatures}
}
\value{
object of class Eigensystem
}
\description{
Set whichEigenfeatures for object of EigensystemPlotParam
}
\details{
Set whichEigenfeatures for object of EigensystemPlotParam
}
\author{
Anneleen Daemen \email{daemen.anneleen@gene.com}, Matthew Brauer \email{brauer.matthew@gene.com}
}
\seealso{
EigensystemPlotParam-class
}

