% Generated by roxygen2 (4.0.1): do not edit by hand
\name{whichFeatures<-}
\alias{whichFeatures<-}
\title{Set whichFeatures for object of EigensystemPlotParam}
\usage{
whichFeatures(x) <- value
}
\arguments{
\item{x}{object of class Eigensystem}

\item{value}{new value for slot whichFeatures}
}
\value{
object of class Eigensystem
}
\description{
Set whichFeatures for object of EigensystemPlotParam
}
\details{
Set whichFeatures for object of EigensystemPlotParam
}
\author{
Anneleen Daemen \email{daemen.anneleen@gene.com}, Matthew Brauer \email{brauer.matthew@gene.com}
}
\seealso{
EigensystemPlotParam-class
}

