% Generated by roxygen2 (4.0.1): do not edit by hand
\name{whichPolarAxes<-}
\alias{whichPolarAxes<-}
\title{Set whichPolarAxes for object of EigensystemPlotParam}
\usage{
whichPolarAxes(x) <- value
}
\arguments{
\item{x}{object of class Eigensystem}

\item{value}{new value for slot whichPolarAxes}
}
\value{
object of class Eigensystem
}
\description{
Set whichPolarAxes for object of EigensystemPlotParam
}
\details{
Set whichPolarAxes for object of EigensystemPlotParam
}
\author{
Anneleen Daemen \email{daemen.anneleen@gene.com}, Matthew Brauer \email{brauer.matthew@gene.com}
}
\seealso{
EigensystemPlotParam-class
}

