% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plot,Eigensystem,EigensystemPlotParam-method}
\alias{plot}
\alias{plot,Eigensystem,EigensystemPlotParam-method}
\title{Generate visualizations of the eigensystem}
\usage{
\S4method{plot}{Eigensystem,EigensystemPlotParam}(x, y, ...)
}
\arguments{
\item{x}{object of class Eigensystem}

\item{y}{object of class EigensystemPlotParam}

\item{...}{Additional arguments that can be passed on}
}
\description{
Generates up to ten visualizations of the eigensystem to aid in deciding which eigenfeatures and eigenassays to filter out (representing noise, steady state, steady-scale, experimental artifacts), or to aid in exploring the dynamics of expression/intensity levels over time or between different groups of assays.
}
\details{
The function generates a heatmap of the eigenfeatures by assays with use of the given contrast factor (eigenfeatureHeatmap), a heatmap of the features by eigenassays with use of the given contrast factor (eigenassayHeatmap), a heatmap of the features by assays, with features sorted according to two selected eigenfeatures (sortedHeatmap), a bar plot with the eigenexpression fractions of all eigenfeatures (fraction), a screeplot for the eigenexpression fractions (scree), a bar plot with the eigenexpression fractions of the eigenfeatures without the dominant eigenfeature(s) (zoomedFraction), the intensity levels of selected eigenfeatures across the assays (by default eigenfeatures 1-4) (lines), the intensity levels of all eigenfeatures across the assays (allLines), polar plot for the assays according to their correlation with two eigenassays (eigenassayPolar), and polar plot for the features according to their correlation with two eigenfeatures (eigenfeaturePolar).
}
\section{Methods}{

\describe{
\item{\code{signature(x = "Eigensystem", y = "EigensystemPlotParam")}}{
}
}
}
\examples{
## Metabolomics starvation data obtained from http://genomics-pubs.princeton.edu/StarvationMetabolomics/Download.shtml
data(StarvationData)

## Computes the eigensystem for the actual data
eigensystem <- compute(StarvationData)
## Generates all provided plots for the eigensystem
params <- new("EigensystemPlotParam")
if (.Platform$OS.type != "windows") plot(eigensystem, params)
## Generates all provided plots for the eigensystem, with use of contrast 2 for the heatmap
contrast(params) <- 2
if (.Platform$OS.type != "windows") plot(eigensystem, params)
## Generates the fraction and lines plot for the eigensystem
params <- new("EigensystemPlotParam")
plots(params) <- c("fraction","lines")
if (.Platform$OS.type != "windows") plot(eigensystem, params)
}
\author{
Anneleen Daemen \email{daemen.anneleen@gene.com}, Matthew Brauer \email{brauer.matthew@gene.com}
}
\references{
Alter O, Brown PO and Botstein D. Singular value decomposition for genome-wide expression data processing and modeling. Proc Natl Acad Sci U.S.A. 97(18), 10101-10106 (2000).
}
\seealso{
Other X.eigensystem.: \code{\link{compute}},
  \code{\link{compute,Eigensystem-method}},
  \code{\link{compute,ExpressionSet-method}},
  \code{\link{compute,data.frame-method}},
  \code{\link{compute,matrix-method}};
  \code{\link{exclude}},
  \code{\link{exclude,Eigensystem-method}};
  \code{\link{report}},
  \code{\link{report,Eigensystem,EigensystemPlotParam-method}}
}

