% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{report,Eigensystem,EigensystemPlotParam-method}
\alias{report}
\alias{report,Eigensystem,EigensystemPlotParam-method}
\title{Creates a report of the eigensystem}
\usage{
\S4method{report}{Eigensystem,EigensystemPlotParam}(x, y, ...)
}
\arguments{
\item{x}{object of class Eigensystem}

\item{y}{object of class EigensystemPlotParam}

\item{...}{Additional arguments that can be passed on}
}
\description{
Generates a summarizing report of the eigensystem
}
\details{
The function generates a txt file of the eigensystem, containing the list of features with their coordinates, radius and phase in the polar plot according to two selected eigenfeatures.
}
\section{Methods}{

\describe{
\item{\code{signature(x = "Eigensystem", y = "EigensystemPlotParam")}}{
}
}
}
\examples{
## Metabolomics starvation data obtained from http://genomics-pubs.princeton.edu/StarvationMetabolomics/Download.shtml
data(StarvationData)

## Computes the eigensystem for the actual data
eigensystem <- compute(StarvationData)
## Exclude the eigenfeatures representing steady-state intensity
eigensystem <- exclude(eigensystem)
## Computes the eigensystem on the variance in the data after filtering out stead-state intensity
eigensystem <- compute(eigensystem, apply="variance")
## No exclusion of eigenfeatures representing steady-scale variance
eigensystem <- exclude(eigensystem, excludeEigenfeatures=0)

## Generate report for eigenfeatures 1 and 2
params <- new("EigensystemPlotParam")
if (.Platform$OS.type \%in\% "windows") path(params) <- getwd()
report(eigensystem, params)
## Generate report for eigenfeatures 2 and 3
whichPolarAxes(params) <- c(3,2)
report(eigensystem, params)
}
\author{
Anneleen Daemen \email{daemen.anneleen@gene.com}, Matthew Brauer \email{brauer.matthew@gene.com}
}
\seealso{
Other X.eigensystem.: \code{\link{compute}},
  \code{\link{compute,Eigensystem-method}},
  \code{\link{compute,ExpressionSet-method}},
  \code{\link{compute,data.frame-method}},
  \code{\link{compute,matrix-method}};
  \code{\link{exclude}},
  \code{\link{exclude,Eigensystem-method}};
  \code{\link{plot}},
  \code{\link{plot,Eigensystem,EigensystemPlotParam-method}}
}

